Clicking on event button now adds and retracts from database
Clicking on event attend button now will add an entry to the user_event database or take away from it, depending on if an entry is present or not.

Changed file sturcture to get rid of error in server about js code not being loaded in properly.

Changed database to be compatable with SQLAlchemy libs.