When getting events.html, buttons set depending on state of DB
An event on events.html, when user is authenticated, will be set to a specific state depending on the presence of items in the database.

There is a bug where the button is duplicated; however, this does not break the application, and is fine being in the application for the time being.